+++ src/moco.py
@@ -130,7 +130,7 @@ class MoCo(nn.Module):
         iter_stats[f"{stats_prefix}loss"] = (loss.item(), bsz)
 
         predicted_idx = torch.argmax(logits, dim=-1)
-        accuracy = 100 * (predicted_idx == labels).float().mean()
+        accuracy = 100 * (predicted_idx == labels).to(torch.bfloat16).mean()
         stdq = torch.std(q, dim=0).mean().item()
         stdk = torch.std(k, dim=0).mean().item()
         iter_stats[f"{stats_prefix}accuracy"] = (accuracy, bsz)
